azure_scripts/requirements.txt
# Python dependencies for Azure SQL Database connection testing
python-dotenv>=1.0.0
pyodbc>=5.0.0
